refactor(web): move payplug notification to payment controller
